+++ testing/e2e/suites/utils.go
@@ -41,7 +41,7 @@ func waitForAvailable(ctx context.Context, c client.Client, deployment appsv1.De
 func upsert(ctx context.Context, c client.Client, obj client.Object) error {
 	copy := obj.DeepCopyObject().(client.Object)
 	lgr := logger.FromContext(ctx).With("object", copy.GetName(), "namespace", copy.GetNamespace())
-	lgr.Info(fmt.Sprintf("upserting object: %s", obj.GetName()))
+	lgr.Info("upserting object")
 
 	// create or update the object
 	lgr.Info("attempting to create object")
